requirements.txt
Pillow
fpdf2
python-dateutil
tkcalendar
num2words
pdf2image
PyMuPDF